# Conflicts:
#	deeppavlov/configs/go_bot/gobot_dstc2_all.json
#	deeppavlov/core/common/chainer.py
#	deeppavlov/core/data/utils.py
#	deeppavlov/core/layers/tf_layers.py
#	deeppavlov/models/embedders/fasttext_embedder.py
#	deeppavlov/models/embedders/glove_embedder.py
#	deeppavlov/skills/go_bot/bot.py
#	requirements.txt

+++ deeppavlov/core/data/utils.py
@@ -17,7 +17,6 @@ limitations under the License.
 from pathlib import Path
 
 import requests
-import sys
 from tqdm import tqdm
 import tarfile
 import gzip
@@ -36,30 +35,35 @@ _MARK_DONE = '.done'
 tqdm.monitor_interval = 0
 
 
-def download(dest_file_path, source_url):
+def download(dest_file_path, source_url, force_download=True):
     """Download a file from URL
 
     Args:
         dest_file_path: path to the file destination file (including file name)
         source_url: the source URL
+        force_download: download file if it already exists, or not
 
     """
     CHUNK = 16 * 1024
     dest_file_path = Path(dest_file_path).absolute()
-    dest_file_path.parent.mkdir(parents=True, exist_ok=True)
 
-    r = requests.get(source_url, stream=True)
-    total_length = int(r.headers.get('content-length', 0))
+    if force_download or not dest_file_path.exists():
+        dest_file_path.parent.mkdir(parents=True, exist_ok=True)
 
-    with dest_file_path.open('wb') as f:
-        log.info('Downloading from {} to {}'.format(source_url, dest_file_path))
+        r = requests.get(source_url, stream=True)
+        total_length = int(r.headers.get('content-length', 0))
 
-        pbar = tqdm(total=total_length, unit='B', unit_scale=True)
-        for chunk in r.iter_content(chunk_size=CHUNK):
-            if chunk:  # filter out keep-alive new chunks
-                pbar.update(len(chunk))
-                f.write(chunk)
-        f.close()
+        with dest_file_path.open('wb') as f:
+            log.info('Downloading from {} to {}'.format(source_url, dest_file_path))
+
+            pbar = tqdm(total=total_length, unit='B', unit_scale=True)
+            for chunk in r.iter_content(chunk_size=CHUNK):
+                if chunk:  # filter out keep-alive new chunks
+                    pbar.update(len(chunk))
+                    f.write(chunk)
+            f.close()
+    else:
+        log.info('File already exists in {}'.format(dest_file_path))
 
 
 def untar(file_path, extract_folder=None):
